+++ src/Resource.php
@@ -5,6 +5,7 @@ namespace Epayco;
 
 class Resource extends Client
 {
+    protected $epayco;
     /**
      * Instance epayco class
      * @param array $epayco
